!htp Merge AuthenticationDirective into SecurityDirectives, refactor to more directive-centric style
